pyproject.toml
[tool.poetry]
name = "atro-args"
version = "0.2.0"
description = ""
authors = ["atropos <sv7n@pm.me>"]
readme = "README.md"
packages = [{include = "atro_args"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "*"
python-dotenv = "*"
pyyaml = "*"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
